# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_photo_api_photos_upload_post": {
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "file": {
            "format": "binary",
            "title": "File",
            "type": "string"
          },
          "tags": {
            "default": "",
            "title": "Tags",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_photo_api_photos_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ItemLinkResolved": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "from_item_id": {
            "title": "From Item Id",
            "type": "string"
          },
          "link_id": {
            "title": "Link Id",
            "type": "string"
          },
          "link_type": {
            "title": "Link Type",
            "type": "string"
          },
          "other_item": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ItemSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_item_id": {
            "title": "To Item Id",
            "type": "string"
          }
        },
        "required": [
          "link_id",
          "from_item_id",
          "to_item_id",
          "link_type",
          "created_at"
        ],
        "title": "ItemLinkResolved",
        "type": "object"
      },
      "ItemLinksResponse": {
        "additionalProperties": false,
        "properties": {
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/ItemLinkResolved"
            },
            "title": "Links",
            "type": "array"
          }
        },
        "required": [
          "item_id"
        ],
        "title": "ItemLinksResponse",
        "type": "object"
      },
...
      "LogbookEntryCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "problem": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Problem",
            "type": "string"
          },
          "related_item_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Item Ids",
            "type": "array"
          },
          "root_cause": {
            "default": "",
            "maxLength": 8000,
            "title": "Root Cause",
            "type": "string"
          },
          "solution": {
            "maxLength": 12000,
            "minLength": 1,
            "title": "Solution",
            "type": "string"
          },
          "source_ref": {
            "default": "",
            "maxLength": 2000,
            "title": "Source Ref",
            "type": "string"
          },
          "source_type": {
            "default": "manual",
            "enum": [
              "manual",
              "document-derived",
              "autotest-derived"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "reviewed",
              "verified",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "default": "",
            "maxLength": 2000,
            "title": "Tags",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "problem",
          "solution"
        ],
        "title": "LogbookEntryCreateRequest",
        "type": "object"
      },
      "LogbookEntryPageResponse": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/LogbookEntryResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "LogbookEntryPageResponse",
        "type": "object"
      },
      "LogbookEntryResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "problem": {
            "title": "Problem",
            "type": "string"
          },
          "related_item_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Item Ids",
            "type": "array"
          },
          "root_cause": {
            "title": "Root Cause",
            "type": "string"
          },
          "run_id": {
            "default": "",
            "title": "Run Id",
            "type": "string"
          },
          "solution": {
            "title": "Solution",
            "type": "string"
          },
          "source_ref": {
            "default": "",
            "title": "Source Ref",
            "type": "string"
          },
          "source_type": {
            "default": "manual",
            "enum": [
              "manual",
              "document-derived",
              "autotest-derived"
            ],
            "title": "Source Type",
            "type": "string"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "reviewed",
              "verified",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "title": "Tags",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "problem",
          "root_cause",
          "solution",
          "tags",
          "created_at",
          "updated_at"
        ],
        "title": "LogbookEntryResponse",
        "type": "object"
      },
      "LogbookEntryUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "problem": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem"
          },
          "related_item_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Item Ids"
          },
          "root_cause": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Cause"
          },
          "solution": {
            "anyOf": [
              {
                "maxLength": 12000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Solution"
          },
          "source_ref": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "source_type": {
            "anyOf": [
              {
                "enum": [
                  "manual",
                  "document-derived",
                  "autotest-derived"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "draft",
                  "reviewed",
                  "verified",
                  "archived"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "tags": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "LogbookEntryUpdateRequest",
        "type": "object"
      },
...
      "MessageResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "MessageResponse",
        "type": "object"
      },
      "PhotoPageResponse": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PhotoResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "title": "PhotoPageResponse",
        "type": "object"
      },
...
      "PhotoUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 8000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "draft",
                  "reviewed",
                  "verified",
                  "archived"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "tags": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "title": "PhotoUpdateRequest",
        "type": "object"
      },
      "PromoteToKnowledgeResponse": {
        "additionalProperties": false,
        "properties": {
          "knowledge_entry_id": {
            "title": "Knowledge Entry Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message",
          "knowledge_entry_id"
        ],
        "title": "PromoteToKnowledgeResponse",
        "type": "object"
      },
...
      "UploadPhotoResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "file_size": {
            "title": "File Size",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ocr_error": {
            "default": "",
            "title": "Ocr Error",
            "type": "string"
          },
          "ocr_status": {
            "default": "pending",
            "enum": [
              "pending",
              "completed",
              "failed",
              "unavailable"
            ],
            "title": "Ocr Status",
            "type": "string"
          },
          "ocr_text": {
            "title": "Ocr Text",
            "type": "string"
          },
          "status": {
            "default": "reviewed",
            "enum": [
              "draft",
              "reviewed",
              "verified",
              "archived"
            ],
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "title": "Tags",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "uploaded_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By"
          }
        },
        "required": [
          "id",
          "filename",
          "tags",
          "description",
          "created_at",
          "updated_at",
          "file_size",
          "ocr_text",
          "message"
        ],
        "title": "UploadPhotoResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Knowledge Workspace API",
    "version": "5.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/logbook/entries": {
      "get": {
        "operationId": "list_logbook_entries_api_logbook_entries_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogbookEntryPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Logbook Entries"
      },
      "post": {
        "operationId": "create_logbook_entry_api_logbook_entries_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogbookEntryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Logbook Entry"
      }
    },
    "/api/logbook/entries/{entry_id}": {
      "delete": {
        "operationId": "delete_logbook_entry_api_logbook_entries__entry_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Logbook Entry"
      },
      "patch": {
        "operationId": "update_logbook_entry_api_logbook_entries__entry_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogbookEntryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Logbook Entry"
      }
    },
    "/api/logbook/entries/{entry_id}/promote-to-knowledge": {
      "post": {
        "operationId": "promote_logbook_to_knowledge_api_logbook_entries__entry_id__promote_to_knowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoteToKnowledgeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Promote Logbook To Knowledge"
      }
    },
...
    "/api/photos": {
      "get": {
        "operationId": "list_photos_api_photos_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Photos"
      }
    },
    "/api/photos/upload": {
      "post": {
        "operationId": "upload_photo_api_photos_upload_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_photo_api_photos_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadPhotoResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Photo"
      }
    },
    "/api/photos/{photo_id}": {
      "delete": {
        "operationId": "delete_photo_api_photos__photo_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "title": "Photo Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Photo"
      },
      "patch": {
        "operationId": "update_photo_api_photos__photo_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "title": "Photo Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhotoUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Photo"
      }
    },
    "/api/photos/{photo_id}/download": {
      "get": {
        "operationId": "download_photo_api_photos__photo_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "title": "Photo Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "inline",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Inline",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Photo"
      }
    },
    "/api/photos/{photo_id}/references": {
      "get": {
        "operationId": "list_photo_references_api_photos__photo_id__references_get",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "title": "Photo Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemLinksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Photo References"
      }
    },
...
  }
}